JAMES-1670 ToRecipientFolder should match our coding conventions
